Exclude owner from assignable users vocabulary.
The main use case concerns the notification feature which uses the assignable users vocabulary. Imagine a member of the workspace creates some content and the user is then removed from the workspace. Calling the notification view on the user's content would still list the user as a possible recipient of the notification because he still is the owner of the content. This is no longer the case with this change.