Fix tests by adding ftw.calendar version constraint.
ftw.workspace depends on ftw.meeting, which depends on ftw.calendar with a version constraint.
ftw.meeting also depends directly on ftw.calendar without a version constraint.
Since buildout does not have an intelligent dependency resolution this can lead to the
problem that ftw.calendar was already picked in an incompatible version.

In order to fix that, we must provide a version pinning on buildout level.